requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pydantic==2.4.2
elasticsearch==8.10.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
numpy==1.26.2
pandas==2.1.3 